backend/requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
faiss-cpu
firebase-admin
python-multipart
pydantic
Pillow
numpy